Route Flogger logs to log4j in connector
Add routing of Flogger logs in Scylla CDC Source Connector to log4j.
Flogger is used in the underlying scylla-cdc-java library and before
this change those logs were not forwarded to log4j. Debezium (as well as
Kafka and Kafka Connect) use the log4j library.

To perform this routing, a dependency on flogger-log4j-backend is added.
In static block `flogger.backend_factory` property is changed to use
the log4j backend. This static block is placed before any other static
initializations and it is present in two classes that can be loaded
by Kafka Connect - ScyllaConnector and ScyllaConnectorTask. Such
placement is very important, as it MUST be executed before any
`FluentLogger.forEnclosingClass()` invocations (which read this System
property).