Implemented record simplified new state extractor
This extractor transforms record in a similar way to
io.debezium.transforms.ExtractNewRecordState, but does not
differentiate between NULL and no value. Thanks to this, values
don't need to be packed in a 1-field struct, and can be represented
directly instead. This transformer makes the connector compatibile
with Elasticsearch Sink - quick setup is documented in
`QUICKSTART-ELASTICSEARCH-INTEGRATION.md` file.