RFC - ScyllaExtractNewRecordState - Dynamically flatten null values
ScyllaExtractNewRecordState payload can't distinguish between NULL
updates and DELETEs, in turn making it impossible for a consumer or
subscriber to reconciliate deltas and converge to a stable state.

This proposal addresses the problem in two ways:

First, it ensures that every consumed message gets its own schema. This
is done via the new createSchemaKey method, which computes a unique key
accordingly to the values sent by the producer.

Second, it drops all null Struct fields via the isDroppableField method,
so that consumers will only observe null values representing affected
columns as observed within an update or delete operation.

This has been tested with both JSON and Avro converters, and the same
set of CQL statements as in https://github.com/scylladb/scylla-cdc-source-connector/issues/64

Refs: https://github.com/scylladb/scylla-cdc-source-connector/issues/64