emitDeleteRecord - Switch to after struct
A "raw" Debezium message contains two Fields within its payload: after
and before. In the RDBMs world, a DELETE message often represents a
`null` after and a pre-image of the affected row within the before Field.

This semantics, however, makes no sense in ScyllaDB. A pre-image is only
possible in a delete operation during a ROW_DELETE operation, and all
other types of deletes (including range deletes) - today - can not be
represented by preImages.

Arguably, representing delete operations in the after field won't make
sense either, since a delete operation by definition should always
result in a null after payload.

In other words, we live in the worst of the two words: We are a NoSQL
database using a format originally thought for relational systems.

The motivation for this commit is quite simple: We need to come up with
a standardized way to handle delete events, including eventually
supporting complex delete events.

The Cassandra connector follows the same approach, and adds additional
metadata to delete operations, such that a specialized consumer can
interpret the delete in question.

After this change, the following structure is defined:

 - The before field is used exclusively by preImages.
 - The after field is used exclusively by deltas or postImages.
 - Supported deletes should always emit an after message.