ScyllaExtractNewState - Generate Dynamic Schemas
ScyllaExtractNewRecordState payload can't distinguish between NULL
updates and DELETEs, in turn making it impossible for a consumer or
subscriber to reconciliate deltas and converge to a stable state.

This commit addresses the problem in two ways:

First, it ensures that every consumed message gets its own schema. This
is done via the new getSchemaKey method, which computes a unique
hashCode accordingly to the values sent by the producer.

Second, it drops all null Struct fields via the isDroppableField method,
so that consumers will only observe null values representing affected
columns as affected observed within an update or delete operation.

Since deletes are now part of the after Field, we also handle the case
when delete.handling.mode is a rewrite and set the respective boolean
flag to true, so that consumers relying on it can reconciliate the
changes.

When we decide to more complex delete types, care should be taken not to
emit wrong tombstones and correctly handling the rewrite mode. This is
beyond the scope of this commit.