Port to Debezium 2.6.2.Final
Some adjustments needed to be made to use the new API.
Main difference is that now there are many classes parametrized by
Partition type and Offset type.
Connector already had 2 implementations of OffsetContext but it's not
possible to have singular Partition type for both of them that makes
sense in all needed contexts. Not every new component is expected to
work flawlessly, but functionality should remain the same.
Snapshotting is not expected to work.
Among changes that can be noticed in usage are the following:
- `scylla.name` was replaced with `topic.prefix` option
- There are additional timestamps in produced messages
- Logs printed by `SchemaNameAdjuster` will point to its package